modify laserTurret

+++ Assets/Scipts/EnemyMove.cs
@@ -6,8 +6,8 @@ using UnityEngine.UI;
 public class EnemyMove : MonoBehaviour {
 
     public float speed = 10;
-    private int totalHp;
-    public int hp = 150;
+    private float totalHp;
+    public float hp = 150;
     private Transform[] positions;//拐点集合
     private int index = 0;
     public GameObject explosionEffect;
@@ -44,7 +44,7 @@ public class EnemyMove : MonoBehaviour {
     {
         EnemySpawner.countEnemyAlive--;
     }
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (hp <= 0) return;
         hp -= damage;
